refactor casting data types

+++ 2_CastingDataTypes.sol
@@ -2,11 +2,11 @@
 
 // Declare the versions of the solidity compiler
 pragma solidity >=0.7.0 <0.9.0;
- 
+
 // Used to convert uint to string
 // import "@openzeppelin/contracts/utils/Strings/sol";
 
 contract CastingDataTypesContract {
-  uint toBig = 250;
-  uint8 castedValue = uint8(toBig);
+    uint256 toBig = 250;
+    uint8 castedValue = uint8(toBig);
 }
